Move hooks and their test files.
Moved oidc client and signal hooks to proper subfolders.
